Yeni simgeler eklendi ve README güncellendi; kullanılmayan JSON dosyaları kaldırıldı ve yeni kullanışlı özellikler eklendi. Buglar düzeltildi ve daha çok kullanılabilir oldu.
