--- Classes/Sample Class/dates/2025-3-4.json
@@ -1,1 +1,0 @@
-{"12345678": "-", "12345679": "+", "12245678": "-"}
