--- Classes/Sample Class/namelist.json
@@ -1,1 +1,0 @@
-{"12345678": "Sample Student 1", "12345679": "Sample Student 2", "12245678": "Sample Student 3"}
