--- Classes/Sample Class/pointlist.json
@@ -1,1 +1,0 @@
-{"12345678": "0", "12345679": "4", "12245678": "0"}
